"exception handling bug fixed"

+++ core/routers/users.py
@@ -1,14 +1,7 @@
 from fastapi import APIRouter,status,HTTPException,Request,Depends
-from fastapi.responses import HTMLResponse,RedirectResponse
 from core import schemas,utils,database,models
 from sqlalchemy.orm import Session
-from core.exception.custom_exceptions import InvalidCredential
 from core.o2auth import create_Access_token,verify_token,get_current_user
-from fastapi.security import OAuth2PasswordRequestForm
-from fastapi.templating import Jinja2Templates 
-
-templates=Jinja2Templates(directory="templates")
-
 router=APIRouter(prefix="/user",tags=["User"])
 
 #home page of user 
@@ -18,12 +11,12 @@ def user_home_page():
 
 @router.get('/profile/')
 def user_profile(request:Request,db:Session=Depends(database.get_db),user_id=Depends(get_current_user)):
-    # user_id=get_current_user(request.cookies.get("token"))
     data=db.get(models.Auth,user_id)
     data.resident   # using lazy loading concept yaad rakhna      
     data.personal
     data.vehicle
-    return {"Auth":data}
+    data.complaint
+    return {"Data":data}
 
 @router.post('/personal/')
 def update_personal(user_data:schemas.Personal,user_id=Depends(get_current_user),db:Session=Depends(database.get_db)):
@@ -60,18 +53,10 @@ def add_vehicle(user_data:schemas.Vehicle,user_id=Depends(get_current_user),db:S
 @router.post('/vehicle/remove/')
 def delete_vehicle(user_data:schemas.Vehicle,user_id=Depends(get_current_user),db:Session=Depends(database.get_db)):
     obj=db.get(models.Vehicle,(user_id,user_data.number))
+    if obj==None:
+        return {'status':False,"messege":f"vehicle does not exit of number: {user_data.number}"}
     db.delete(obj)
     db.commit()
     return {'status':True,"messege":"vehicle delete  succcesfully"}
-x=[]
-@router.post('/complaint')
-def submit_complaiant(user_data:schemas.Complaint,db:Session=Depends(database.get_db),user=Depends(get_current_user)):
-    obj=models.Complaint(user_id=user,category=user_data.category,description=user_data.description,attachement=user_data.attachement,subject=user_data.subject)
-    obj=obj.__dict__
-    x.append(obj)
-    return {"status":True,"messege":"Complaint submit succesfully!","data":obj}
 
-@router.get('/complaint')
-def query_status(user_id=Depends(get_current_user),db:Session=Depends(database.get_db)):
-    # data=db.query(models.Complaint).filter(models.Complaint.user_id==user_id).all()
-    return {"status":True,"data":x}
+
